js/xpconnect/tests/mochitest/test_bug1005806.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1005806
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1005806</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1005806">Mozilla Bug 1005806</a>
<p id="display"></p>
<div id="content" style="display: none">
  <input id="ipt"></input>
</div>
<pre id="test">
</pre>
<script type="application/javascript">

/** Test for Bug 1005806 **/
is(typeof document.getElementById('ipt').controllers, 'undefined', "Controllers property should not appear for content");

</script>
</body>
</html>
